✨ Code Architecture, Readme and Changelog
