+++ .eslintrc.cjs
@@ -1,14 +1,10 @@
 module.exports = {
-  env: { browser: true, es2020: true },
-  extends: [
-    'eslint:recommended',
-    'plugin:@typescript-eslint/recommended',
-    'plugin:react-hooks/recommended',
-  ],
-  parser: '@typescript-eslint/parser',
-  parserOptions: { ecmaVersion: 'latest', sourceType: 'module' },
-  plugins: ['react-refresh'],
-  rules: {
-    'react-refresh/only-export-components': 'warn',
-  },
-}
+	env: { browser: true, es2020: true },
+	extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended', 'plugin:react-hooks/recommended'],
+	parser: '@typescript-eslint/parser',
+	parserOptions: { ecmaVersion: 'latest', sourceType: 'module' },
+	plugins: ['react-refresh'],
+	rules: {
+		'react-refresh/only-export-components': 'warn',
+	},
+};
