+++ .prettierrc.json
@@ -1,0 +1,9 @@
+{
+	"trailingComma": "es5",
+	"semi": true,
+	"singleQuote": true,
+	"printWidth": 120,
+	"tabWidth": 4,
+	"useTabs": true,
+	"endOfLine": "auto"
+}
