+++ CHANGELOG.MD
@@ -1,0 +1,14 @@
+## 0.0.2 (Released - 07-06-2023)
+
+_Features:_
+
+-   Code Architecture
+-   Readme and Changelog
+
+## 0.0.1 (Released - 06-06-2023)
+
+_Features:_
+
+-   Repo Init
+-   Zustand Store
+-   API initialization
