+++ src/assets/images/index.ts
@@ -1,0 +1,5 @@
+import Pokeball from './pokeball.png';
+
+export const images = {
+	Pokeball,
+};
