+++ src/services/pokemonServices.ts
@@ -1,17 +1,15 @@
-import { AbilityDetail, BaseList, Pokemon } from "../types/pokemon";
-import { httpCommon } from "../utils/http-common";
-import axios from "axios";
+import { AbilityDetail, BaseList, Pokemon } from '../types/pokemon';
+import { httpCommon } from '../utils/http-common';
+import axios from 'axios';
 
-export const getPokemonList = async ({
-  pageParam = "https://pokeapi.co/api/v2/pokemon",
-}) => {
-  return axios.get<BaseList>(pageParam);
+export const getPokemonList = async ({ pageParam = 'https://pokeapi.co/api/v2/pokemon' }) => {
+	return axios.get<BaseList>(pageParam);
 };
 
 export const getPokemon = async (name: string) => {
-  return httpCommon.get<Pokemon>(`/pokemon/${name}`);
+	return httpCommon.get<Pokemon>(`/pokemon/${name}`);
 };
 
 export const getPokemonAbility = async (ability: string) => {
-  return await httpCommon.get<AbilityDetail>(`/ability/${ability}`);
+	return await httpCommon.get<AbilityDetail>(`/ability/${ability}`);
 };
