+++ src/utils/http-common.ts
@@ -1,5 +1,5 @@
-import axios from "axios";
+import axios from 'axios';
 
 export const httpCommon = axios.create({
-  baseURL: "https://pokeapi.co/api/v2",
+	baseURL: 'https://pokeapi.co/api/v2',
 });
