+++ tailwind.config.js
@@ -1,9 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 export default {
-  content: ['./src/**/*.tsx', './index.html'],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
-
+	content: ['./src/**/*.tsx', './index.html'],
+	theme: {
+		extend: {},
+	},
+	plugins: [],
+};
